chore: remove replaced bot command handlers, reimplement Sus as feature, cleanup
