--- src/ShitpostBot/src/ShitpostBot.Application/Features/BotCommands/BotCommand.cs
@@ -1,5 +1,0 @@
-namespace ShitpostBot.Application.Features.BotCommands;
-
-public record BotCommand(string Command);
-
-public record BotCommandEdit(ulong BotResponseMessageId);
