+++ src/ShitpostBot/src/ShitpostBot.Application/Features/PostTracking/MessageDeleted.cs
@@ -1,0 +1,6 @@
+using MediatR;
+using ShitpostBot.Infrastructure;
+
+namespace ShitpostBot.Application.Features.PostTracking;
+
+public record MessageDeleted(MessageIdentification Identification) : INotification;
